BlogAPIBackend/env_04_04_2023/requirements.txt
flask==2.1.3
flask-smorest
python-dotenv
SQLAlchemy==1.4
flask-sqlalchemy
flask-jwt-extended
passlib
flask-migrate
gunicorn
psycopg2
requests
redis
rq
Flask-Cors==3.0.10
flask-crontab==0.1.2
Flask-Caching==2.0.2